Refactoring to use proper variable naming
