+++ config/Shared/config_default.php
@@ -593,8 +593,8 @@ $config[FileSystemConstants::FILESYSTEM_SERVICE]['s3-import'] = [
     'sprykerAdapterClass' => Aws3v3FilesystemBuilderPlugin::class,
     'root' => '/',
     'path' => '/',
-    'key' => getenv('SPRYKER_S3_CSV_BUCKET_KEY') ?: '',
-    'secret' => getenv('SPRYKER_S3_CSV_BUCKET_SECRET') ?: '',
+    'key' => getenv('DATA_IMPORT_S3_KEY') ?: '',
+    'secret' => getenv('DATA_IMPORT_S3_SECRET') ?: '',
     'bucket' => 'tarxter-prod-csv-uploads',
     'version' => 'latest',
     'region' => 'eu-central-1',
